refactor(section):removed section folder into components folder/remove duplicate landing page entry point
Deleted HomeLanding.js as its functionality was redundant with index.js.
Consolidated the landing page entry point to use index.js for better maintainability and to eliminate code duplication./Removed section folder into components folder
