+++ package.json
@@ -14,9 +14,8 @@
     "@emotion/styled": "^11.14.0",
     "@mui/icons-material": "^7.0.2",
     "@mui/material": "^7.0.2",
-    "@radix-ui/react-slot": "^1.2.2",
-    "@supabase/ssr": "^0.6.1",
     "@radix-ui/react-separator": "^1.1.6",
+    "@radix-ui/react-slot": "^1.2.2",
     "@radix-ui/react-tabs": "^1.1.11",
     "@supabase/ssr": "^0.6.1",
     "@supabase/supabase-js": "^2.49.4",
