+++ src/components/sections/BannerByCinema/BannerByCinema.js
