--- src/pages/home-landing/HomeLanding.js
@@ -1,21 +1,0 @@
-import React from "react";
-import BannerByCinema from "./sections/BannerByCinema/BannerByCinema";
-import FrameByCinema from "./sections/FrameByCinema/FrameByCinema";
-import NavbarByCinema from "./sections/NavbarByCinema/NavbarByCinema";
-import FooterSection from "./sections/FooterSection/FooterSection";
-import Navbar from "@/components/Navbar/Navbar";
-
-const HomeLandingDefault = () => {
-  return (
-    <div className="relative w-full">
-      <div className="w-full">
-        <Navbar />
-        <BannerByCinema />
-      </div>
-      <FrameByCinema />
-      <FooterSection />
-    </div>
-  );
-};
-
-export default HomeLandingDefault;
