+++ src/pages/home-landing/index.js
@@ -1,11 +1,9 @@
 import React, { useState } from 'react';
-import NavbarByCinema from "./sections/NavbarByCinema/NavbarByCinema";
-import BannerByCinema from "./sections/BannerByCinema/BannerByCinema";
+import Navbar from "@/components/Navbar/Navbar";
+import BannerByCinema from "../../components/sections/BannerByCinema/BannerByCinema";
 import AnnouncementPopup from "../../components/AnnouncementPopup";
 import FrameByCinema from "./sections/FrameByCinema/FrameByCinema";
 import FooterSection from "./sections/FooterSection/FooterSection";
-import Head from 'next/head';
-import Navbar from '@/components/Navbar/Navbar';
 
 const HomeLanding = () => {
   const [filters, setFilters] = useState({
@@ -26,8 +24,8 @@ const HomeLanding = () => {
     <main className="min-h-screen bg-background overflow-x-hidden">
       <AnnouncementPopup />
       <NavbarByCinema />
-      <BannerByCinema onSearch={handleSearch} />
-      <FrameByCinema filters={searchFilters} />
+      <BannerByCinema />
+      <FrameByCinema />
       <FooterSection />
     </main>
   );
